add Abstract Fabric

+++ src/com/company/Figures/FabricRectangle.java
@@ -2,7 +2,7 @@ package com.company.Figures;
 
 public class FabricRectangle {
 
-    public static Rectangle create(int lenght, int width){
+    public static Rectangle create(double lenght, double width){
 
         if( lenght > 0 && width > 0) {
             return new Rectangle(lenght, width);
