+++ src/com/company/Figures/Triangle.java
@@ -4,14 +4,14 @@ import java.util.Objects;
 
 public class Triangle implements IFigure {
 
-    private int sideA;
-    private int sideB;
-    private int sideC;
+    private double sideA;
+    private double sideB;
+    private double sideC;
 
     public Triangle() {
     }
 
-    public Triangle(int sideA, int sideB, int sideC) {
+    public Triangle(double sideA, double sideB, double sideC) {
         this.sideA = sideA;
         this.sideB = sideB;
         this.sideC = sideC;
@@ -37,27 +37,27 @@ public class Triangle implements IFigure {
         return r;
     }
 
-    public int getSideA() {
+    public double getSideA() {
         return sideA;
     }
 
-    public void setSideA(int sideA) {
+    public void setSideA(double sideA) {
         this.sideA = sideA;
     }
 
-    public int getSideB() {
+    public double getSideB() {
         return sideB;
     }
 
-    public void setSideB(int sideB) {
+    public void setSideB(double sideB) {
         this.sideB = sideB;
     }
 
-    public int getSideC() {
+    public double getSideC() {
         return sideC;
     }
 
-    public void setSideC(int sideC) {
+    public void setSideC(double sideC) {
         this.sideC = sideC;
     }
